Add tests and fix issues

+++ app/Http/Controllers/Api/User/Favorite/DeleteController.php
@@ -4,17 +4,24 @@ namespace App\Http\Controllers\Api\User\Favorite;
 
 use App\Http\Controllers\Controller;
 use App\Repositories\EloquentFavoriteRepository;
-use Illuminate\Http\JsonResponse;
+use Illuminate\Auth\Access\AuthorizationException;
+use Illuminate\Http\Response;
 
 class DeleteController extends Controller
 {
     /**
      * Handle the incoming request.
+     *
+     * @throws AuthorizationException
      */
-    public function __invoke($id, EloquentFavoriteRepository $favorites): JsonResponse
+    public function __invoke($id, EloquentFavoriteRepository $favorites): Response
     {
+        $favorite = $favorites->find($id);
+
+        $this->authorize('delete', $favorite);
+
         $favorites->delete($id);
 
-        return response()->json(['message' => 'Favorite deleted successfully']);
+        return response()->noContent();
     }
 }
